requirements.txt
# ── Core ──────────────────────────────────────────────────────────────
python-dotenv>=1.0.0       # Environment variable loading

# ── Scheduling ────────────────────────────────────────────────────────
schedule>=1.2.0             # Intraday task scheduler

# ── Data / Analysis ───────────────────────────────────────────────────
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0

# ── Market Data Feeds ─────────────────────────────────────────────────
yfinance>=0.2.40           # Yahoo Finance — global + NSE quotes (live)
# nsepython>=2.0.0         # NSE live data (optional, uncomment if needed)

# ── Machine Learning (Strategy Evolution) ─────────────────────────────
scikit-learn>=1.3.0

# ── Broker SDKs (install the one you use) ─────────────────────────────
# Zerodha KiteConnect
kiteconnect>=4.2.0

# Dhan HQ
dhanhq>=2.0.0

# AngelOne SmartAPI
smartapi-python>=1.3.7
pyotp>=2.9.0              # TOTP for AngelOne 2FA

# ── HTTP / WebSocket ──────────────────────────────────────────────────
requests>=2.31.0
websocket-client>=1.6.0

# ── Control Tower Dashboard ──────────────────────────────────────────────
streamlit>=1.32.0

# ── Utilities ─────────────────────────────────────────────────────────
loguru>=0.7.0              # Optional: richer logging
tqdm>=4.66.0               # Progress bars for evolution / backtests
